Move custom reporters to separate file
